+++ docs/src/game/button.config.ts
@@ -1,0 +1,5 @@
+interface ButtonConfig {
+    button: HTMLButtonElement;
+    visible: boolean;
+    disabled: boolean;
+}