+++ docs/src/game/game.js
@@ -42,13 +42,19 @@ export class Game {
             this.betInput.min = this.betInputMin;
             this.betInput.value = this.betInput.min;
             this.betInput.max = this.player.money.toString();
-            this.updateButtons(false, false, [this.hitButton, this.standButton, this.doubleDownButton]);
-            this.updateButtons(true, false, [this.placeBetButton]);
+            this.updateButtons([
+                { button: this.hitButton, visible: false, disabled: true },
+                { button: this.standButton, visible: false, disabled: true },
+                { button: this.placeBetButton, visible: true, disabled: true },
+                { button: this.doubleDownButton, visible: false, disabled: true }
+            ]);
             yield Promise.all([
                 this.dealer.clearHand(),
                 this.player.clearHand()
             ]);
-            this.updateButtons(true, true, [this.placeBetButton]);
+            this.updateButtons([
+                { button: this.placeBetButton, visible: true, disabled: false }
+            ]);
             this.placeBetButton.focus();
         });
     }
@@ -60,8 +66,11 @@ export class Game {
     }
     hit() {
         this.ripple(this.hitButton);
-        const buttons = [this.hitButton, this.standButton];
-        this.updateButtons(true, false, [...buttons, this.doubleDownButton]);
+        this.updateButtons([
+            { button: this.hitButton, visible: true, disabled: true },
+            { button: this.standButton, visible: true, disabled: true },
+            { button: this.doubleDownButton, visible: !this.doubleDownButton.disabled, disabled: true }
+        ]);
         this.player.addCard();
         if (this.player.score > Rules.BLACKJACK) {
             setTimeout(() => {
@@ -70,11 +79,14 @@ export class Game {
             }, 500);
             return;
         }
-        if (this.player.score === Rules.BLACKJACK) {
+        if ((this.player.score === Rules.BLACKJACK)) {
             this.stand();
             return;
         }
-        this.updateButtons(true, true, buttons);
+        this.updateButtons([
+            { button: this.hitButton, visible: true, disabled: false },
+            { button: this.standButton, visible: true, disabled: false }
+        ]);
         this.hitEnterKeyListener();
     }
     ripple(button) {
@@ -96,7 +108,10 @@ export class Game {
     }
     stand() {
         this.ripple(this.standButton);
-        this.updateButtons(true, false, [this.hitButton, this.standButton, this.doubleDownButton]);
+        this.updateButtons([
+            { button: this.hitButton, visible: true, disabled: false },
+            { button: this.standButton, visible: true, disabled: false },
+        ]);
         this.dealer.flipCard();
         setTimeout(() => this.startDealersTurn(), 500);
     }
@@ -111,12 +126,31 @@ export class Game {
         }, this.DEALER_DELAY_MS);
     }
     bet() {
-        this.player.placeBet();
-        this.dealCards();
+        return __awaiter(this, void 0, void 0, function* () {
+            this.hitEnterKeyListener();
+            this.updateButtons([
+                { button: this.placeBetButton, visible: false, disabled: false }
+            ]);
+            if (this.player.hasBlackjack()) {
+                this.updateButtons([
+                    { button: this.hitButton, visible: true, disabled: true },
+                    { button: this.standButton, visible: true, disabled: true },
+                    { button: this.doubleDownButton, visible: true, disabled: true },
+                ]);
+                return yield new Promise(resolve => setTimeout(resolve, 3e3)).then(() => this.stand());
+            }
+            this.updateButtons([
+                { button: this.hitButton, visible: true, disabled: false },
+                { button: this.standButton, visible: true, disabled: false },
+                { button: this.doubleDownButton, visible: true, disabled: !this.player.canDoubleDown },
+            ]);
+            this.player.placeBet();
+            this.dealCards();
+        });
     }
     doubleDown() {
-        // TODO
         this.ripple(this.doubleDownButton);
+        this.hit();
     }
     split() {
         // TODO
@@ -174,17 +208,7 @@ export class Game {
         this.placeBetButton.addEventListener('blur', () => this.placeBetButton.focus());
         this.placeBetButton.addEventListener('focus', () => this.betInput.step = this.betInputMin);
         this.placeBetButton.addEventListener('wheel', (event) => (event.deltaY > 0) ? this.betInput.stepDown() : this.betInput.stepUp());
-        this.placeBetButton.addEventListener('click', () => __awaiter(this, void 0, void 0, function* () {
-            this.hitEnterKeyListener();
-            this.bet();
-            this.updateButtons(false, false, [this.placeBetButton]);
-            const actionButtons = [this.hitButton, this.standButton, this.doubleDownButton];
-            if (this.player.hasBlackjack()) {
-                this.updateButtons(true, false, actionButtons);
-                return yield new Promise(resolve => setTimeout(resolve, 3e3)).then(() => this.stand());
-            }
-            this.updateButtons(true, true, actionButtons);
-        }));
+        this.placeBetButton.addEventListener('click', () => this.bet());
     }
     /**
      * Allow to use the 'Enter' key to hit after placing a bet
@@ -205,10 +229,10 @@ export class Game {
             this.hit();
         });
     }
-    updateButtons(setVisible, setEnabled, buttons) {
-        buttons.forEach(button => {
-            button.style.display = setVisible ? 'block' : 'none';
-            button.disabled = !setEnabled;
+    updateButtons(config) {
+        config.forEach(cfg => {
+            cfg.button.style.display = cfg.visible ? 'block' : 'none';
+            cfg.button.disabled = cfg.disabled;
         });
     }
 }