+++ docs/src/player/player.ts
@@ -30,6 +30,10 @@ export class Player extends Role {
         return !this.blackjack ? this.bet : (this.bet * Game.BLACKJACK_PAYOUT);
     }
 
+    get canDoubleDown() {
+        return (this.bet * 2) <= this.money;
+    }
+
     placeBet(): void {
         const bet = (document.getElementById('bet') as HTMLInputElement).value;